Optimize search by using gridmap and accumulating is_passable
Using cumulative probability distribution from each node's is_passable
seems like a good idea, but in reality it is unnecessarily slow with a
lot of nodes because it has to scan all nodes. Rather, we can try to
randomly pick a node and if it is not is_passable just try again.
It sounds stupid but performs better with a lot of nodes.