Use a combination of line segments and arcs instead of Bezier curves
Bezier curves have more smooth transition of curvature, but it is
inherently difficult to measure the Euclidean distance, so the
algorithms will have hard time calculation correct length and positions
of the rail nodes.

Arcs on the other hand has analytically easier properties and in
particular it can calculate distance exactly.